Macronutriens with consumed foodss are sent to frontend, mealplanDTO is created

+++ src/main/java/com/example/nutry/model/MacroNutrientsDTO.java
@@ -10,7 +10,7 @@ import lombok.*;
 @ToString
 public class MacroNutrientsDTO {
 
-    private Integer protein;
-    private Integer fat;
-    private Integer carbohydrate;
+    private Double protein;
+    private Double fat;
+    private Double carbohydrate;
 }
